Complete overhaul of the internal design of the multi-threaded apps

+++ run_ffmpeg/src/run_ffmpeg/ProbeDirectoriesWorkerThread.java
@@ -13,16 +13,11 @@ import com.mongodb.client.model.Filters;
 /**
  * A worker thread for ProbeDirectories.
  */
-public class ProbeDirectoriesWorkerThread extends Thread
+public class ProbeDirectoriesWorkerThread extends run_ffmpegWorkerThread
 {
 	/// Reference back to the controller thread.
-	private transient ProbeDirectories pdController = null ;
+	private transient ProbeDirectories theController = null ;
 
-	/// Setup the logging subsystem
-	private transient Logger log = null ;
-	private transient Common common = null ;
-
-	//	private transient MoviesAndShowsMongoDB masMDB = null ;
 	private transient MongoCollection< FFmpegProbeResult > probeInfoCollection = null ;
 
 	/// This map will store all of the FFmpegProbeResults in the probeInfoCollection, keyed by the long path to the document.
@@ -41,23 +36,20 @@ public class ProbeDirectoriesWorkerThread extends Thread
 	public ProbeDirectoriesWorkerThread( ProbeDirectories pdController,
 			Logger log,
 			Common common,
-			//			MoviesAndShowsMongoDB masMDB,
 			MongoCollection< FFmpegProbeResult > probeInfoCollection,
 			Map< String, FFmpegProbeResult > probeInfoMap,
 			List< String > foldersToProbe )
 	{
-		assert( pdController != null ) ;
-		assert( log != null ) ;
-		assert( common != null ) ;
-		//		assert( masMDB != null ) ;
+		super( log, common ) ;
+		
+		assert( theController != null ) ;
 		assert( probeInfoCollection != null ) ;
 		assert( probeInfoMap != null ) ;
 		assert( foldersToProbe != null ) ;
 
-		this.pdController = pdController ;
+		this.theController = pdController ;
 		this.log = log ;
 		this.common = common ;
-		//		this.masMDB = masMDB ;
 		this.probeInfoCollection = probeInfoCollection ;
 		this.probeInfoMap = probeInfoMap ;
 		this.foldersToProbe = foldersToProbe ;
@@ -69,20 +61,42 @@ public class ProbeDirectoriesWorkerThread extends Thread
 	 */
 	protected void checkForMissingFiles()
 	{
-		List< File > foundFiles = common.getFilesInDirectoryByExtension( common.addPathSeparatorIfNecessary( getName() ), extensionsToProbe ) ;
-		//		log.info( getName() + " Found " + foundFiles.size() + " file(s)" ) ;
+		List< File > foundFiles = new ArrayList< File >() ;
+
+		// Iterate through each of the folders to probe. For each, find all files in the directory tree.
+		for( String folderToProbe : foldersToProbe )
+		{
+			if( !shouldKeepRunning() )
+			{
+				return ;
+			}
+			List< File > filesInThisFolder = common.getFilesInDirectoryByExtension(
+					folderToProbe, extensionsToProbe ) ;
+			// getFilesInDirectoryByExtension() may return an empty List, but it is guaranteed to be non-null.
+			foundFiles.addAll( filesInThisFolder ) ;
+		}		
+		log.info( getName() + " Found " + foundFiles.size() + " file(s)" ) ;
 
 		// Need to walk through the probeInfoMap and check if each entry corresponds to an active file.
-		// Will need to be able to search the files by path.
+		// Will need to be able to search the files by path -- put the foundFiles into a map.
 		Map< String, File > fileSystemMap = new HashMap< String, File >() ;
 		for( File theFile : foundFiles )
 		{
+			if( !shouldKeepRunning() )
+			{
+				return ;
+			}
 			fileSystemMap.put( theFile.getAbsolutePath(), theFile ) ;
 		}
 
 		// Now walk through the probeInfoMap to search for missing files.
 		for( Map.Entry< String, FFmpegProbeResult > entry : probeInfoMap.entrySet() )
 		{
+			if( !shouldKeepRunning() )
+			{
+				return ;
+			}
+			
 			final String absolutePath = entry.getKey() ;
 			final FFmpegProbeResult theProbeResult = entry.getValue() ;
 
@@ -90,7 +104,7 @@ public class ProbeDirectoriesWorkerThread extends Thread
 			final File theFile = fileSystemMap.get( absolutePath ) ;
 			if( null == theFile )
 			{
-				log.info( getName() + " Deleting missing file: " + absolutePath ) ;
+				log.info( getName() + " Deleting missing file from database: " + absolutePath ) ;
 				if( !common.getTestMode() )
 				{
 					probeInfoCollection.deleteOne( Filters.eq( "_id", theProbeResult._id ) ) ;
@@ -101,7 +115,7 @@ public class ProbeDirectoriesWorkerThread extends Thread
 
 	/**
 	 * Return an FFmpegProbeResult corresponding to the given fileToProbe
-	 * if it exists in the database.
+	 * if it exists in the database (as stored in memory).
 	 * @param probeInfoCollection
 	 * @param fileToProbe
 	 * @return
@@ -142,6 +156,7 @@ public class ProbeDirectoriesWorkerThread extends Thread
 			// Size has changed
 			return true ;
 		}
+
 		if( fileToProbe.lastModified() > probeResult.getLastModified() )
 		{
 			// File has been modified since the last probe
@@ -164,7 +179,7 @@ public class ProbeDirectoriesWorkerThread extends Thread
 		if( !forceRefresh && ((probeResult != null) && !needsRefresh( fileToProbe, probeResult )) )
 		{
 			// No need to probe again.
-			log.fine( "File already exists and does not need a refresh, skipping: " + fileToProbe.getAbsolutePath() ) ;
+			log.fine( getName() + " File already exists and does not need a refresh, skipping: " + fileToProbe.getAbsolutePath() ) ;
 			return probeResult ;
 		}
 		// Post-condition: File does not currently exist in the database, or it does and it needs a refresh, or it's null
@@ -172,18 +187,21 @@ public class ProbeDirectoriesWorkerThread extends Thread
 		if( probeResult != null )
 		{
 			// In the case it needs a refresh, just delete the old one and re-probe
-			log.fine( "Deleting probeResult: " + probeResult ) ;
-			probeInfoCollection.deleteOne( Filters.eq( "_id", probeResult._id ) ) ;
+			log.fine( getName() + " Deleting probeResult: " + probeResult ) ;
+			if( !common.getTestMode() )
+			{
+				probeInfoCollection.deleteOne( Filters.eq( "_id", probeResult._id ) ) ;
+			}
 		}
 
 		// File needs a probe
-		log.fine( "Probing " + fileToProbe.getAbsolutePath() ) ;
+		log.fine( getName() + " Probing " + fileToProbe.getAbsolutePath() ) ;
 
 		// Handle the special case that this is a missing file substitute
 		if( fileToProbe.getName().contains( Common.getMissingFilePreExtension() ) )
 		{
 			// Missing file. Do not probe directly
-			log.fine( "This is a missing file: " + fileToProbe.getAbsolutePath() );
+			log.fine( getName() + " This is a missing file: " + fileToProbe.getAbsolutePath() );
 			probeResult = new FFmpegProbeResult() ;
 			probeResult.setFileNameWithPath( fileToProbe.getAbsolutePath() ) ;
 			probeResult.setFileNameWithoutPath( fileToProbe.getName() ) ;
@@ -203,12 +221,15 @@ public class ProbeDirectoriesWorkerThread extends Thread
 		// Push the probe result into the database.
 		if( probeResult != null )
 		{
-			probeInfoCollection.insertOne( probeResult ) ;
+			if( !common.getTestMode() )
+			{
+				probeInfoCollection.insertOne( probeResult ) ;
+			}
 			probeInfoMap.put( probeResult.getFileNameWithPath(), probeResult ) ;
 		}
 		else
 		{
-			log.warning( "Null probeResult for file " + fileToProbe.getAbsolutePath() ) ;
+			log.warning( getName() + " Null probeResult for file " + fileToProbe.getAbsolutePath() ) ;
 		}
 		return probeResult ;
 	}
@@ -221,7 +242,14 @@ public class ProbeDirectoriesWorkerThread extends Thread
 		// Walk through each folder
 		for( String folderToProbe : foldersToProbe )
 		{
-			log.fine( "Probing folder: " + folderToProbe ) ;
+			if( !shouldKeepRunning() )
+			{
+				// Stop running
+				//				log.info( "Shutting down thread" ) ;
+				break ;
+			}
+
+			log.fine( getName() + " Probing folder: " + folderToProbe ) ;
 
 			// Find files in this folder to probe
 			List< File > filesToProbe = common.getFilesInDirectoryByExtension( folderToProbe, extensionsToProbe ) ;
@@ -233,23 +261,21 @@ public class ProbeDirectoriesWorkerThread extends Thread
 				if( !shouldKeepRunning() )
 				{
 					// Stop running
-					log.info( "Shutting down thread" ) ;
+					//					log.info( getName() + " Shutting down thread" ) ;
 					break ;
 				}
 
 				probeFileAndUpdateDB( fileToProbe ) ;
 			} // for( fileToProbe )
 
-			log.info( "Completed probing folder: " + folderToProbe ) ;
+			log.info( getName() + " Completed probing folder: " + folderToProbe ) ;
 		} // for( folderToProbe )
-
-
 	}
 
 	@Override
 	public void run()
 	{
-		log.info( "Probing drives and folders: " + foldersToProbe.toString() ) ;
+		log.info( getName() + " Probing drives and folders: " + foldersToProbe.toString() ) ;
 
 		{
 			final long startTime = System.nanoTime() ;
@@ -258,7 +284,7 @@ public class ProbeDirectoriesWorkerThread extends Thread
 			checkForMissingFiles() ;
 
 			final long endTime = System.nanoTime() ;
-			log.info( "Finished checking for missing files " + foldersToProbe.toString()
+			log.info( getName() + " Finished checking for missing files " + foldersToProbe.toString()
 			+ ", " + common.makeElapsedTimeString( startTime, endTime ) ) ;
 		}
 
@@ -269,13 +295,13 @@ public class ProbeDirectoriesWorkerThread extends Thread
 			probeNewFiles() ;
 
 			final long endTime = System.nanoTime() ;
-			log.info( "Finished probing " + foldersToProbe.toString()
+			log.info( getName() + " Finished probing " + foldersToProbe.toString()
 			+ ", " + common.makeElapsedTimeString( startTime, endTime ) ) ;
 		}
 	} // run()
-
+	
 	public boolean shouldKeepRunning()
 	{
-		return pdController.shouldKeepRunning() ;
+		return theController.shouldKeepRunning() ;
 	}
 }
